Progress on Debian packaging
One freaking week later, I might have finally figured Debian
packaging out.

The DKMS script (/usr/sbin/dkms) is probably meant only to generate
basic packages for local/unofficial use. This is not only for the
awkward template mechanism that it uses, but also because of the
flags that it runs dpkg-buildpackage with:

	dpkg-buildpackage -rfakeroot -d -b -us -uc

-d, -us and -uc suggest to me that the resulting package is meant
to be a convenience throwaway one.

So I think I'm meant to get rid of the "DKMS script" layer and
issue dpkg-buildpackage directly. That doesn't mean I'm getting rid
of DKMS; only that I'm not using the famed script.

I also discovered that, apparently, I'm meant to create two
separate packages; one for the kernel modules and another one for
the userspace clients. One reason why this is probably the case is
because the kernel modules are supposed to be packaged with
architecture "all" (since it's just a bunch of code that will be
compiled later) while the userspace apps are "any" (since they are
C binaries). There might be other reasons, but I'm not sure.

I have also further improved the standards compliance, but I'm not
done yet. I still have to run lintian, piuparts, check dependencies
in a chroot environment, upgrade the package format to "3 (quilt)"
(I think) and probably other stuff I'm forgetting. I would also
like to figure out why `MAKE[1]` won't work in dkms.conf (and
jool-dkms.dkms).

I'm also apparently supposed to remove the debian directory from
the repository. But whatever; later.