Fix a few dpkg-buildpackage warnings and the make distclean error
Apparently, autotools has this bug where two binaries cannot share
source files, otherwise make distclean errors. Since I don't have
the time to make a library, I had to resort to this workaround
in which `make` doesn't compile the common source directly, but
rather, a bunch of proxy .c files that #include the common ones.
With this, the common .c files are accesed but their directories
don't get polluted with traditional autotools dirt files.

The downside is that it looks needlessly convoluted. Well, it's not
needless at all. It needs to be this way because autotools sucks.
Fixes #264.

Regarding `debian/jool-dkms.install`: The script was improved to
more cleanly copy the right files from `debian/tmp` to
`debian/jool-dkms`.
It hadn't dawned upon me that this file is a script, and I can
therefore use foreaches. Meh. The script is now less sensitive to
directure structure changes the kernel module might suffer in the
future.

There's also the tweaks for `usr/configure.ac`, but that has a
massive comment and I don't want to repeat myself here.

Regarding the package-building, it's still missing:

- joold needs a manpage
- `lintian -EviIL +pedantic` spews a couple of wishlist-severity
  bugs.
- test build in a chroot environment

And I haven't even started reviewing the RPM solution...